Mejoras periodo y catalogo contable

+++ WebApp/AltivaWebApp/DomainsConta/CatalogoContable.cs
@@ -14,16 +14,16 @@ namespace AltivaWebApp.DomainsConta
         public string Notas { get; set; }
         public short Nivel { get; set; }
         public short IdTipoCuentaContable { get; set; }
-        public int IdCuentaContablePadre { get; set; }
+        public long IdCuentaContablePadre { get; set; }
         public string CuentaContablePadre { get; set; }
         public string DescCuentaPadre { get; set; }
         public bool Movimiento { get; set; }
-        public int IdCuentaPresupuesto { get; set; }
+        public long IdCuentaPresupuesto { get; set; }
         public bool Evaluacion { get; set; }
         public int IdMonedaEvaluacion { get; set; }
         public int IdTipoConversion { get; set; }
         public bool Inactivo { get; set; }
-       
+      
 
     }
 }
